feat: impl commandframe & rename package & apiversion
