+++ src/main/java/com/nextplugins/warps/NextWarps.java
@@ -1,14 +1,12 @@
-package com.nextplugins.nextwarps;
+package com.nextplugins.warps;
 
 import com.github.eikefab.libs.minecraft.InventoryRegistry;
-import com.nextplugins.nextwarps.api.warp.WarpFile;
-import com.nextplugins.nextwarps.api.warp.adapter.WarpAdapter;
-import com.nextplugins.nextwarps.api.warp.adapter.WarpItemAdapter;
-import com.nextplugins.nextwarps.commands.WarpCommand;
-import com.nextplugins.nextwarps.configuration.GeneralValue;
-import com.nextplugins.nextwarps.configuration.registry.ConfigurationRegistry;
+import com.nextplugins.warps.api.warp.WarpFile;
+import com.nextplugins.warps.api.warp.adapter.WarpAdapter;
+import com.nextplugins.warps.api.warp.adapter.WarpItemAdapter;
+import com.nextplugins.warps.commands.WarpCommand;
+import com.nextplugins.warps.configuration.registry.ConfigurationRegistry;
 import lombok.Getter;
-import me.bristermitten.pdm.PluginDependencyManager;
 import me.saiintbrisson.bukkit.command.BukkitFrame;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -51,29 +49,18 @@ public final class NextWarps extends JavaPlugin {
 
     @Override
     public void onEnable() {
-        PluginDependencyManager.of(this).loadAllDependencies().thenRun(() -> {
-            inventoryRegistry = InventoryRegistry.of(this);
+        inventoryRegistry = InventoryRegistry.of(this);
 
-            ConfigurationRegistry.of(this).inject();
+        ConfigurationRegistry.of(this).inject();
 
-            bukkitFrame = new BukkitFrame(this);
-            bukkitFrame.registerCommands(new WarpCommand(warpFile));
+        bukkitFrame = new BukkitFrame(this);
+        bukkitFrame.registerCommands(new WarpCommand(warpFile));
 
-            configureBStats();
-        });
+        new Metrics(this, PLUGIN_ID);
     }
 
     private FileConfiguration getFile(String path) {
         return YamlConfiguration.loadConfiguration(new File(getDataFolder(), path));
     }
-
-    private void configureBStats() {
-        if (!GeneralValue.get(GeneralValue::metrics)) return;
-
-        new Metrics(this, PLUGIN_ID);
-
-        this.getLogger().info("Enabled bStats successfully, statistics enabled");
-    }
-
 
 }
