+++ src/main/java/com/nextplugins/warps/api/warp/WarpInventory.java
@@ -1,12 +1,11 @@
-package com.nextplugins.nextwarps.api.warp;
+package com.nextplugins.warps.api.warp;
 
 import com.github.eikefab.libs.minecraft.InventoryBuilder;
 import com.github.eikefab.libs.minecraft.InventoryRegistry;
 import com.github.eikefab.libs.minecraft.inventory.CustomInventory;
-import com.nextplugins.nextwarps.NextWarps;
-import com.nextplugins.nextwarps.api.warp.adapter.WarpItemAdapter;
-import com.nextplugins.nextwarps.configuration.GeneralValue;
-import lombok.Getter;
+import com.nextplugins.warps.NextWarps;
+import com.nextplugins.warps.api.warp.adapter.WarpItemAdapter;
+import com.nextplugins.warps.configuration.GeneralValue;
 import org.bukkit.inventory.Inventory;
 
 public final class WarpInventory {
