+++ src/main/resources/config.yml
@@ -1,6 +1,3 @@
-# Ative ou desative o uso do bStats (apenas serve para coletar algumas informações sobre o uso do plugin como quantidade de pessoas usando, servidores e etc...)
-metrics: true
-
 menu:
   size: 27
   name: "Warps"
@@ -19,6 +16,5 @@ menu:
       lore:
         - '&7Spawn'
       warp: spawn
-
 
 config-version: 1
